requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
python-dotenv==1.0.0  #  до тех пор, пока не запустим в контейнере
pytz==2022.7.1
sqlparse==0.4.3
